requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyMuPDF==1.24.9
easyocr==1.7.2
Pillow==10.4.0
